aibot/requirements.txt
openai
python-telegram-bot